[network_info_plus] adapt the macOS plugin
