+++ packages/network_info_plus_macos/CHANGELOG.md
@@ -1,49 +1,3 @@
-## 0.3.0
-
-- Updated platform interface
-
-## 0.2.0
-
-- Transfer to plus-plugins monorepo
-
-## 0.1.0+6
-
-- Transfer package to Flutter Community under new name `network_info_plus_macos`.
-
-## 0.1.0+5
-
-- Update package:e2e reference to use the local version in the flutter/plugins
-  repository.
-
-## 0.1.0+4
-
-- Remove Android folder from `connectivity_macos`.
-
-## 0.1.0+3
-
-- Bump the minimum Flutter version to 1.12.13+hotfix.5.
-- Clean up various Android workarounds no longer needed after framework v1.12.
-- Complete v2 embedding support.
-- Fix CocoaPods podspec lint warnings.
-- Declare API stability and compatibility with `1.0.0` (more details at: https://github.com/flutter/flutter/wiki/Package-migration-to-1.0.0).
-
-## 0.1.0+2
-
-- Remove hard coded ios workspace setting of the example app.
-
-## 0.1.0+1
-
-- Make the pedantic dev_dependency explicit.
-
 ## 0.1.0
 
-- Adds an example app to trigger CI tests. Bumped the MINOR version to
-  avoid compatibility issues once this packages is endorsed.
-
-## 0.0.2+1
-
-- Add CHANGELOG.
-
-## 0.0.1
-
-- Initial open source release.
+- Initial release.
