+++ packages/network_info_plus_macos/lib/network_info_plus_macos.dart
@@ -1,3 +1,3 @@
 // Analyze will fail if there is no main.dart file. This file should
-// be removed once an example app has been added to connectivity_macos.
+// be removed once an example app has been added to network_info_plus_macos.
 // https://github.com/flutter/flutter/issues/51007
