+++ packages/network_info_plus_macos/macos/Classes/NetworkInfoPlusPlugin.swift
@@ -1,0 +1,48 @@
+// Copyright 2019 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+import Cocoa
+import CoreWLAN
+import FlutterMacOS
+import SystemConfiguration.CaptiveNetwork
+
+public class NetworkInfoPlusPlugin: NSObject, FlutterPlugin {
+  var cwinterface: CWInterface?
+
+  public override init() {
+    cwinterface = CWWiFiClient.shared().interface()
+  }
+
+  public static func register(with registrar: FlutterPluginRegistrar) {
+    let channel = FlutterMethodChannel(
+      name: "dev.fluttercommunity.plus/network_info",
+      binaryMessenger: registrar.messenger)
+
+    let instance = NetworkInfoPlusPlugin()
+    registrar.addMethodCallDelegate(instance, channel: channel)
+  }
+
+  public func handle(_ call: FlutterMethodCall, result: @escaping FlutterResult) {
+    switch call.method {
+    case "wifiName":
+      result(cwinterface?.ssid())
+    case "wifiBSSID":
+      result(cwinterface?.bssid())
+    case "wifiIPAddress":
+      result(getWifiIP())
+    default:
+      result(FlutterMethodNotImplemented)
+    }
+  }
+}
