+++ packages/network_info_plus_macos/macos/network_info_plus_macos.podspec
@@ -4,14 +4,14 @@
 Pod::Spec.new do |s|
   s.name             = 'network_info_plus_macos'
   s.version          = '0.0.1'
-  s.summary          = 'Flutter plugin for checking connectivity'
+  s.summary          = 'Flutter plugin for checking network info'
   s.description      = <<-DESC
-  Desktop implementation of the connectivity plugin
+  Desktop implementation of the network info plugin
                        DESC
-  s.homepage         = 'https://github.com/flutter/plugins/tree/main/packages/connectivity/connectivity_macos'
+  s.homepage         = 'https://github.com/fluttercommunity/plus_plugins'
   s.license          = { :type => 'BSD', :file => '../LICENSE' }
   s.author           = { 'Flutter Team' => 'flutter-dev@googlegroups.com' }
-  s.source           = { :http => 'https://github.com/flutter/plugins/tree/main/packages/connectivity/connectivity_macos' }
+  s.source           = { :http => 'https://github.com/fluttercommunity/plus_plugins/tree/main/packages/network_info_plus_macos' }
   s.source_files     = 'Classes/**/*'
   s.dependency 'FlutterMacOS'
   s.dependency 'Reachability'
